+++ xposed-module/app/src/main/java/com/specter/module/ui/AppPickerActivity.java
@@ -1,4 +1,4 @@
-package com.fleet.idrotate.ui;
+package com.specter.module.ui;
 
 import android.app.Activity;
 import android.content.SharedPreferences;